First commit of files to produdce a PhoSim postage stamp of galxies produced by PhoSim from Anže's catalog.
This version uses a toypsf with a width of 0.7.

You can use it like this if WL is set to the directory where these files live:

(from PhoSim directory)
time ./phosim -e 0 -c $WL/toypsf $WL/one_line.txt

(from this directory)
ipython postageStamp.py

It will read the file produced by PhoSim and write a new poststamp.fits file.

The files and part are hardcoded to the source in one_line_catalog.txt
now.  The magnitude of this file has also been articificially
increased to make it more visible.